requirements.txt
PyPDF2
openai
python-dotenv
pdfplumber
colorama
